refactor: Remove legacy modification dropdown
