+++ operate/client/src/App/ProcessInstance/TopPanel/ModificationDropdown/index.tsx
@@ -24,7 +24,10 @@ import {
   Button,
   InlineLoading,
 } from './styled';
-import {getSelectedRunningInstanceCount} from 'modules/utils/flowNodeSelection';
+import {
+  clearSelection,
+  getSelectedRunningInstanceCount,
+} from 'modules/utils/flowNodeSelection';
 import {
   useTotalRunningInstancesByFlowNode,
   useTotalRunningInstancesForFlowNode,
@@ -44,11 +47,15 @@ import {useProcessDefinitionKeyContext} from 'App/Processes/ListView/processDefi
 import {useProcessInstanceXml} from 'modules/queries/processDefinitions/useProcessInstanceXml';
 import {getFlowNodeName} from 'modules/utils/flowNodes';
 import {getParentElement} from 'modules/bpmn-js/utils/getParentElement';
-import {processInstanceDetailsStore} from 'modules/stores/processInstanceDetails';
+import {
+  useIsRootNodeSelected,
+  useRootNode,
+} from 'modules/hooks/flowNodeSelection';
+import {useProcessInstance} from 'modules/queries/processInstance/useProcessInstance';
 
 type Props = {
   selectedFlowNodeRef?: SVGSVGElement;
-  diagramCanvasRef?: React.RefObject<HTMLDivElement>;
+  diagramCanvasRef?: React.RefObject<HTMLDivElement | null>;
 };
 
 const ModificationDropdown: React.FC<Props> = observer(
@@ -64,14 +71,17 @@ const ModificationDropdown: React.FC<Props> = observer(
       useTotalRunningInstancesVisibleForFlowNode(flowNodeId);
     const {data: totalRunningInstancesByFlowNode} =
       useTotalRunningInstancesByFlowNode();
+    const isRootNodeSelected = useIsRootNodeSelected();
     const selectedRunningInstanceCount = getSelectedRunningInstanceCount({
       totalRunningInstancesForFlowNode: totalRunningInstances || 0,
-      isRootNodeSelected: flowNodeSelectionStore.isRootNodeSelected,
+      isRootNodeSelected,
     });
     const availableModifications = useAvailableModifications(
       selectedRunningInstanceCount,
     );
     const canBeModified = useCanBeModified();
+    const rootNode = useRootNode();
+    const {data: processInstance} = useProcessInstance();
 
     const processDefinitionKey = useProcessDefinitionKeyContext();
     const {data: processDefinitionData} = useProcessInstanceXml({
@@ -164,14 +174,13 @@ const ModificationDropdown: React.FC<Props> = observer(
                                   parentScopeIds: generateParentScopeIds(
                                     businessObjects,
                                     flowNodeId,
-                                    processInstanceDetailsStore?.state
-                                      .processInstance?.bpmnProcessId,
+                                    processInstance?.processDefinitionId,
                                   ),
                                 },
                               });
                             }
 
-                            flowNodeSelectionStore.clearSelection();
+                            clearSelection(rootNode);
                           }}
                         >
                           Add
@@ -197,7 +206,7 @@ const ModificationDropdown: React.FC<Props> = observer(
                               flowNodeInstanceId,
                               businessObjects,
                             );
-                            flowNodeSelectionStore.clearSelection();
+                            clearSelection(rootNode);
                           }}
                         >
                           Cancel instance
@@ -223,7 +232,7 @@ const ModificationDropdown: React.FC<Props> = observer(
                               totalRunningInstancesVisible ?? 0,
                               businessObjects,
                             );
-                            flowNodeSelectionStore.clearSelection();
+                            clearSelection(rootNode);
                           }}
                         >
                           Cancel all
@@ -243,7 +252,7 @@ const ModificationDropdown: React.FC<Props> = observer(
                               flowNodeId,
                               flowNodeInstanceId,
                             );
-                            flowNodeSelectionStore.clearSelection();
+                            clearSelection(rootNode);
                           }}
                         >
                           Move instance
@@ -257,7 +266,7 @@ const ModificationDropdown: React.FC<Props> = observer(
                         renderIcon={ArrowRight}
                         onClick={() => {
                           modificationsStore.startMovingToken(flowNodeId);
-                          flowNodeSelectionStore.clearSelection();
+                          clearSelection(rootNode);
                         }}
                       >
                         Move all
